Various updates to code when using 'this' keyword
In various places in the code the 'this' keyword was used. This
was incorrect and would cause this to be undefined. A variable
was established in functions within the Player class constructor
where the 'this' keyword was required. Errors related to use of
'this' are now fixed.

Removed playerInstances functions and integrated them into the
Player class constructor.

Various syntax and missing spaces were also tidied up.